+++ db/migrate/20130228141410_add_import_info_to_designs.rb
@@ -1,0 +1,6 @@
+class AddImportInfoToDesigns < ActiveRecord::Migration
+  def change
+    add_column :designs, :rows_imported, :integer, null: false, default: 0
+    add_column :designs, :total_rows, :integer, null: false, default: 0
+  end
+end