+++ db/migrate/20130228150558_add_import_times_to_design.rb
@@ -1,0 +1,6 @@
+class AddImportTimesToDesign < ActiveRecord::Migration
+  def change
+    add_column :designs, :import_started_at, :datetime
+    add_column :designs, :import_ended_at, :datetime
+  end
+end