+++ test/controllers/user_mailer_test.rb
@@ -109,4 +109,15 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/The data export you requested for #{export.project.name} is now ready for download\./, email.encoded)
   end
 
+  test "import complete email" do
+    design = designs(:one)
+
+    email = UserMailer.import_complete(design).deliver
+    assert !ActionMailer::Base.deliveries.empty?
+
+    assert_equal [design.user.email], email.to
+    assert_equal "Your Design Data Import for #{design.project.name} is Complete", email.subject
+    assert_match(/The design data import for #{design.project.name} is now complete\./, email.encoded)
+  end
+
 end