+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Enhancements
 - **Design Changes**
   - Designs along with data can now be imported from a CSV
+    - Design imports run as background tasks that notify the user by email when the design import is finished
+    - Design import progress can be seen on the design page
 - **Gem Changes**
   - Updated to Rails 4.0.0.beta1
   - Updated to Contour 2.0.0.beta.1